create dashboard.interface
